Fix NaN issue in the pitch angle calculation
The B value for the trap center that corresponds to the initial pitch
angle was calculated as the first B value in time. For use with the
signal simulation with proper retarded time the first retarded time is
most likely not t=0, hence that B value is not the right one.