Modify CRESanaModel inputs
n_samples is not a fixed input anymore. Instead it defaults to None and
users have to explicitly set a value using the setter function. While
set to None the call method raises an exception. This is supposed to
enable loading a predefined model but adjusting n_samples based on track
length.